ajout parametre stepsize et constantstepsize dans les algorithmes (palm4MSA et hierarchical_fact et donc les fichiers de config xml)

+++ test/src/launch_hierarchical_fact.cpp
@@ -8,7 +8,7 @@
 #include "faust_core.h"
 
 
-typedef float FPP;
+typedef double FPP;
 using namespace std;
 
 int main(int argc, char* argv[])
